Extract pullrequest module
